Move CLI commands back into the cli package
Most of the code in Command()s (now Hub() and Agent()) are cli related. Only the construction
and calling of the Agent and Hub objects are specific to their subdomains.

Some of the logic around config files can probably be pushed back down into hub.
